splitting MatchEditor (instructions/positions and part of Handler functions)
