+++ client/lamby-app/src/Instructions_&_Positions/PositionCodes.tsx
@@ -1,0 +1,17 @@
+import { ICode} from '../interfaces/ICode';
+
+const positionCodes: ICode[] = [
+  { code: 'GK', content: 'Goalkeeper' },
+  { code: 'LB', content: 'Left Back' },
+  { code: 'LCB', content: 'Left Center Back' },
+  { code: 'RCB', content: 'Right Center Back' },
+  { code: 'RB', content: 'Right Back' },
+  { code: 'LDM', content: 'Left Defensive Mid' },
+  { code: 'RDM', content: 'Right Defensive Mid' },
+  { code: 'CAM', content: 'Central Attacking Mid' },
+  { code: 'LW', content: 'Left Winger' },
+  { code: 'ST', content: 'Striker' },
+  { code: 'RW', content: 'Right Winger' },
+];
+
+export default positionCodes;
