+++ client/lamby-app/src/interfaces/ICode.ts
@@ -1,0 +1,4 @@
+export interface ICode {
+  code: string;
+  content: string;
+}
